Table value parameter to get around parameter limit
