+++ source/Nevermore/UpgradeScripts/Script0003-RelatedDocumentType.sql
@@ -1,0 +1,1 @@
+﻿CREATE TYPE dbo.RelatedDocumentTableValuedParameter AS TABLE ([Id] NVARCHAR(400), [Table] NVARCHAR(400), [RelatedDocumentId] NVARCHAR(400), [RelatedDocumentTable] NVARCHAR(400))
